Raise APIError if GitHub returns bad status code.
If GitHub.com can't handle the API request, either due to user error or server
error, it returns an utterly useless chunk of HTML. This confuses the library,
so we now raise an APIError if the status code is anything other than 200.
Better error handling will have to wait until the API supports it.