Improved handling of API errors.
GitHub API errors aren't reported consistently yet. Sometimes they return
reams of HTML and a non-200 status code, other times they return a 200 status
code and an error message. We now handle both of these cases by raising an
APIError with an appropriate message. This will need to be updated as the API
standardises. When it does, we want to interpret 404s, for example, as the
object not being found, and thus provide an informative error message.