Tweaked Repository.find_all's argument handling.
Repository.find_all accepted an array of 'words', which it concatenated with
'+'. It now also accepts a single space-separated String, or any combination
of the two.

This method is still buggy, however, in that the query is not URI escaped;
it's simply interpolated as-is into the URI. Peculiarly, the API doesn't
accept URI-escaped space-separated queries, e.g. 'ruby%20spec'. This is
non-standard enough to put off escaping until I know exactly what the API
expects.