Add credentials to `default_params` for auth'd GET.
We want the token and login to be sent for all authenticated queries. They
were being sent for POST requests, but, seemingly, not for GETs, causing
methods relying on the latter to fail. HTTParty's `default_params` method
causes parameters so set to be sent on every request. We specify `login` and
`token` as default parameters if the request is authenticated.