Make .property a class method for Resources.
So as to prevent methods using routes other than the object's default from
having to suffer the performance penalty of the main route being called first,
we want such methods to be implemented as class methods. Methods like
User#followers is an example. It uses the current .property method which is an
instance method. This change will allow such methods the desired flexibility.